v1.21.90: rework delay strategy by resending ContainerOpenPacket until client acknowledges the request Mojang now handles duplicate ContainerOpenPackets gracefully by not resending the window. There is no proper way to know whether the client has indeed rendered a container, but we can infer that by (ab)using PacketViolationWarningPacket.
It is important to 'know' on our end if a container was successfully sent so we know when to terminate
the retry logic. Since a few versions ago, you can no longer view a container whilst having chat window
opened. In these cases not only do you not get to see your inventory, but all your inventories get
bricked.

You can reproduce this bug in vanilla PocketMine as well - press T soon as you right-click a chest. You
can now no longer use E to access your inventory. This commit also takes care of it situation - InvMenu
inventories will not be a source of the issue.