v4.7.0: A new delay strategy to reduce excessive wait time. The delay strategy has been reworked. InvMenu will send excess ContainerOpenPackets until client acknowledges the request. Mojang now handles duplicate ContainerOpenPackets gracefully by not resending the window. While there is no proper way to know whether the client has indeed rendered a container, we can infer this by (ab)using PacketViolationWarningPacket.
Knowing whether a container was successfully sent is important for servers. It is not just so we know when to terminate our fancy retry logic. From a few versions ago, clients can no longer view a container whilst having chat window opened. Otherwise not only do you not get to see your inventory, but all your inventories get "bricked" (inaccessible).

You can reproduce the bug in vanilla PocketMine as well - press T soon as you right-click a chest. Now you can no longer use E to access your inventory. Changes in this commit also takes care of it so InvMenu will not be a source of the issue.